Implement file upload

+++ modules/v1/controllers/ArticleController.php
@@ -4,10 +4,13 @@ namespace notes\modules\v1\controllers;
 
 use notes\components\BehaviorsTrait;
 use notes\modules\v1\models\Article;
+use yii\helpers\FileHelper;
 use yii\helpers\Url;
 use yii\rest\Controller;
+use yii\web\BadRequestHttpException;
 use yii\web\NotFoundHttpException;
 use yii\web\ServerErrorHttpException;
+use yii\web\UploadedFile;
 
 class ArticleController extends Controller
 {
@@ -37,6 +40,9 @@ class ArticleController extends Controller
     public function actionDelete(int $id)
     {
         $model = Article::findOne($id);
+        if (!$model) {
+            throw new NotFoundHttpException("Article $id not found");
+        }
         if ($model->delete() === false) {
             throw new ServerErrorHttpException('Failed to delete article for unknown reason.');
         }
@@ -87,8 +93,58 @@ class ArticleController extends Controller
         return Article::findPopularItems();
     }
 
-    // TODO implement upload action
     public function actionUpload()
     {
+        $userId = 99;
+
+        $file = UploadedFile::getInstanceByName('file');
+
+        if (empty($file)) {
+            throw new BadRequestHttpException('File not uploaded');
+        }
+
+        $fileEndings = [
+            'image/jpeg' => 'jpg',
+            'image/jpg' => 'jpg',
+            'image/gif' => 'gif',
+            'image/png' => 'png'
+        ];
+
+        if ($file->error > 0) {
+            throw new BadRequestHttpException('An error occured');
+        }
+
+        if (!in_array($file->type, array_keys($fileEndings))) {
+            throw new BadRequestHttpException('Invalid file type');
+        }
+
+        if ($file->size > 1000000) {
+            throw new BadRequestHttpException('File size to big');
+        }
+
+        $basename = md5_file($file->tempName);
+        if ($basename === false) {
+            throw new ServerErrorHttpException('Could not create md5 sum');
+        }
+
+        // determine path and url
+        $relPath = sprintf('media/%s/%s.%s', $userId, $basename, $fileEndings[$file->type]);
+        $absFilePath = \Yii::getAlias('@webroot/' . $relPath);
+        $fileUrl = \Yii::getAlias('@web/' . $relPath);
+
+        // create dir if not exists
+        $directory = dirname($absFilePath);
+        if (!is_dir($directory)) {
+            FileHelper::createDirectory($directory, 0775, true);
+        }
+
+        if (!move_uploaded_file($file->tempName, $absFilePath)) {
+            throw new ServerErrorHttpException('Could not move uploaded file');
+        }
+
+        return [
+            'name' => $file->name,
+            'location' => $fileUrl
+        ];
     }
 }
